mystical_python/requirements.txt
# Mystical Python Integration Dependencies
# Core HTTP and API communication
requests>=2.31.0
websocket-client>=1.7.0

# Text processing and fuzzy matching
fuzzywuzzy[speedup]>=0.18.0
python-levenshtein>=0.21.1

# Data processing and validation
pydantic>=2.5.0
numpy>=1.24.0
pyyaml>=6.0.1

# Image and visualization
matplotlib>=3.7.0
pillow>=10.0.0

# Optional scientific packages (for sacred geometry)
scipy>=1.11.0

# Development and testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0